backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
flask-pymongo==2.3.0
python-dotenv==1.0.0
pyotp==2.8.0
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
Pillow==10.0.0
qrcode==7.4.2